pyproject.toml
[project]
name = "content"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "uvicorn==0.30.0",
  "fastapi==0.112.0",
  "langchain-openai==0.3.27",
  "openai>=1.86.0,<2.0.0",
  "langchain-community==0.3.26",
  "langchain-core>=0.3.13,<0.4.0",
  "python-dotenv==1.0.1",
  "httpx==0.27.0",
  "accelerate==1.8.1",
  "beautifulsoup4==4.12.3",
  "bs4>=0.0.2",
  "docx2txt==0.9",
  "faiss-cpu==1.11.0",
  "huggingface>=0.0.1",
  "ipykernel>=6.29.5",
  "langchain==0.3.26",
  "langchain-chroma==0.2.4",
  "langchain-cli>=0.0.36",
  "langchain-huggingface==0.1.2",
  "langgraph>=0.5.0",
  "prompt-template>=1.2.0",
  "protobuf==3.20.0",
  "pydantic>=2.7.4,<3.0.0",
  "pypdf==4.3.1",
  "python-multipart==0.0.5",
  "requests==2.32.3",
  "sentence-transformers==3.1.1",
  "streamlit==1.39.0",
  "transformers==4.53.0",
  "unstructured==0.18.2",
  "jq>=1.9.1",
  "textstat==0.7.3",
]
